+++ src/main/java/heimerdinger/HeimerdingerException.java
@@ -1,3 +1,5 @@
+package heimerdinger;
+
 public class HeimerdingerException extends Exception {
     public HeimerdingerException(String message) {
         super(message);
