+++ src/main/java/heimerdinger/command/AddDeadlineCommand.java
@@ -1,5 +1,10 @@
+package heimerdinger.command;
+
+import heimerdinger.*;
+import heimerdinger.task.Deadline;
+import heimerdinger.task.TaskList;
+
 import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
 public class AddDeadlineCommand extends Command {
