+++ src/main/java/heimerdinger/command/AddToDoCommand.java
@@ -1,3 +1,10 @@
+package heimerdinger.command;
+
+import heimerdinger.*;
+import heimerdinger.task.Task;
+import heimerdinger.task.TaskList;
+import heimerdinger.task.ToDo;
+
 public class AddToDoCommand extends Command {
     private final String description;
 
@@ -7,7 +14,7 @@ public class AddToDoCommand extends Command {
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws HeimerdingerException {
         if (description.isEmpty()) {
-            throw new HeimerdingerException("I can't read your mind great scientist. ToDo must have an accompanying text!");
+            throw new HeimerdingerException("I can't read your mind great scientist. Heimerdinger.ToDo must have an accompanying text!");
         }
         Task t = new ToDo(description);
         tasks.add(t);
