+++ src/main/java/heimerdinger/task/Task.java
@@ -1,3 +1,5 @@
+package heimerdinger.task;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
