+++ src/main/java/heimerdinger/task/ToDo.java
@@ -1,3 +1,5 @@
+package heimerdinger.task;
+
 public class ToDo extends Task {
     String icon = "T";
